+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added `ThroughSuperuserModelBackend`
 * Make list of local IP's for ``get_client_ip`` configurable
 * Added `lock` function that uses `FileLocker` or `CacheLocker` to lock multiple script executions
+* Added post migrate hook to create `view_*` permissions for all content types
 
 ## 0.3.0